requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
requests
